Fourth task fourth try

+++ src/main/java/lang/print/gaps/task4/SettingTypes.java
@@ -7,7 +7,7 @@ char        second = 'a';
 boolean        third = 5 > 9;
 double        forth = 12.90434323324342;
 float        fifth = 12.3F;
-int        sixth = 1;
+short        sixth = 1;
 short        seventh = 129;
     }
 }
